PMM-7922 rename LabelWrapper to Label
