+++ public/app/percona/shared/components/Form/AsyncSelectField/AsyncSelectField.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { AsyncSelect } from '@grafana/ui';
-import { LabelWrapper } from '../LabelWrapper';
+import { Label } from '../Label';
 import { AsyncSelectField } from './AsyncSelectField';
 
 describe('AsyncSelectField', () => {
   it('should render', () => {
     const wrapper = mount(<AsyncSelectField label="label" name="name" onChange={jest.fn()} />);
-    expect(wrapper.find(LabelWrapper).exists()).toBeTruthy();
+    expect(wrapper.find(Label).exists()).toBeTruthy();
     expect(wrapper.find(AsyncSelect).exists()).toBeTruthy();
   });
 });
