+++ public/app/percona/shared/components/Form/AsyncSelectField/AsyncSelectField.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import { AsyncSelect } from '@grafana/ui';
 import { withSelectStyles } from '../withSelectStyles/withSelectStyles';
-import { LabelWrapper } from '../LabelWrapper';
+import { Label } from '../Label';
 import { AsyncSelectFieldProps } from './AsyncSelectField.types';
 
 const AsyncSelectFieldWrapper: FC<AsyncSelectFieldProps<any>> = ({ label, name, ...props }) => (
   <>
-    <LabelWrapper label={label} dataQa={`${name}-select-label`} />
+    <Label label={label} dataQa={`${name}-select-label`} />
     <AsyncSelect {...props} />
   </>
 );
