+++ public/app/percona/shared/components/Form/Label/Label.types.ts
@@ -1,4 +1,4 @@
-export interface FieldWrapperProps {
+export interface LabelProps {
   label: string;
   dataQa?: string;
 }
