+++ public/app/percona/shared/components/Form/Label/index.tsx
@@ -1,0 +1,1 @@
+export * from './Label';
