--- public/app/percona/shared/components/Form/LabelWrapper/index.tsx
@@ -1,1 +1,0 @@
-export * from './LabelWrapper';
