+++ public/app/percona/shared/components/Form/MultiSelectField/MultiSelectField.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { MultiSelect } from '@grafana/ui';
-import { LabelWrapper } from '../LabelWrapper';
+import { Label } from '../Label';
 import { MultiSelectField } from './MultiSelectField';
 
 describe('MultiSelectField', () => {
   it('should render', () => {
     const wrapper = mount(<MultiSelectField label="label" name="name" onChange={jest.fn()} />);
-    expect(wrapper.find(LabelWrapper).exists()).toBeTruthy();
+    expect(wrapper.find(Label).exists()).toBeTruthy();
     expect(wrapper.find(MultiSelect).exists()).toBeTruthy();
   });
 });
