+++ public/app/percona/shared/components/Form/MultiSelectField/MultiSelectField.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { MultiSelect } from '@grafana/ui';
 import { MultiSelectCommonProps } from '@grafana/ui/src/components/Select/types';
 import { withSelectStyles } from '../withSelectStyles/withSelectStyles';
-import { LabelWrapper } from '../LabelWrapper';
+import { Label } from '../Label';
 import { MultiSelectFieldProps } from './MultiSelectField.types';
 
 const MultiSelectFieldWrapper: FC<MultiSelectFieldProps & MultiSelectCommonProps<any>> = ({
@@ -11,7 +11,7 @@ const MultiSelectFieldWrapper: FC<MultiSelectFieldProps & MultiSelectCommonProps
   ...props
 }) => (
   <>
-    <LabelWrapper label={label} dataQa={`${name}-select-label`} />
+    <Label label={label} dataQa={`${name}-select-label`} />
     <MultiSelect {...props} />
   </>
 );
